move homework add screen to quest screen
